+++ src/query/mod.rs
@@ -2,3 +2,4 @@ pub mod query;
 pub mod selectors;
 
 mod traversal;
+mod traversal_composites;
